requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
mss>=9.0.1
scikit-image>=0.19.0
pynput>=1.7.6
Pillow>=9.0.0
matplotlib>=3.5.0 